requirements.txt
numpy
Pillow
tqdm
python-dateutil